Typecast buf_lru_flush_page_count and mem_daptive_hash RHS into ulong

+++ storage/innobase/buf/buf0flu.cc
@@ -1660,7 +1660,7 @@ buf_flush_LRU_list_batch(
 	/* We keep track of all flushes happening as part of LRU
 	flush. When estimating the desired rate at which flush_list
 	should be flushed, we factor in this value. */
-	buf_lru_flush_page_count += n->flushed;
+	buf_lru_flush_page_count += (ulong) n->flushed;
 
 	ut_ad(buf_pool_mutex_own(buf_pool));
